requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==12.24.0
loguru==0.7.2
python-dotenv==1.0.1
scikit-learn==1.4.2
scipy==1.14.1
sentence-transformers==3.3.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
xgboost==2.1.3
numpy==2.1.3
pandas==2.2.3
torcheval==0.0.7
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
slack_sdk==3.33.5
